services/backend/requirements.txt
django
whitenoise
django-cors-headers
django-health-check
djangorestframework
djangorestframework_simplejwt
python-dotenv
psycopg2-binary
gunicorn
django-rest-passwordreset
hashids
django-elasticsearch-dsl
elasticsearch-dsl
elasticsearch